Refactor auth tests
